Выделен модуль для отрисовки задачи
